+++ app/layout.tsx
@@ -1,4 +1,3 @@
-// 'use client'
 import "./globals.css";
 import Footer from "@/components/Footer";
 import Navbar from "@/components/Navbar";
