+++ components/Footer.tsx
@@ -1,25 +1,7 @@
 import React from "react";
 import Image from "next/image";
-import Link from "next/link";
 import { footerLinks } from "@/constant";
-
-type ColumnProps = {
-  title: string;
-  links: Array<string>;
-};
-
-const FooterColumn = ({ title, links }: ColumnProps) => (
-  <div className="footer_column">
-    <h4 className="font-semibold">{title}</h4>
-    <ul className="flex flex-col gap-2 font-normal">
-      {links.map((link) => (
-        <Link href="/" key={link}>
-          {link}
-        </Link>
-      ))}
-    </ul>
-  </div>
-);
+import FooterColumn from "@/section/FooterColumn";
 
 const Footer = () => {
   return (
